Move `get_hyperparameters_space()` function to `class AAIAutogluonTask`
Functionality very similar for both regression and classification, so the same function can be used for both tasks. Function is inherited from `class AAIAutogluonTask` by both regression and classification tasks. Avoids code duplication.
